Add files via upload
Wikipedia is a free online encyclopaedia that is written and edited by volunteers from around the world. It is one of the most popular websites in the world, with over 500 million unique visitors per month. Even though it can be edited by anybody, not everybody edits it. Wikipedia has often been criticized for its lack of diversity. There is a significant gender gap in the number of editors, with only 13.7% of editors being women (Meneses and Poblete 2019). This imbalance in contributions and therefore content is yet another case of the masculine culture of technoscience (Heather Ford and Judy Wajcman, 2017). This gender bias can lead to content gaps and a skewed representation of topics, affecting the inclusivity and accuracy of Wikipedia's content. This can lead to further negative consequences: first, it means that Wikipedia is not as representative of the world as it could be; second, it can make it difficult for women to find information about themselves and their interests on Wikipedia and third, it can discourage women from participating in Wikipedia. There are a number of reasons why there is a gender gap in Wikipedia editors. One reason could be that women are less likely than men to have the technical skills needed to edit Wikipedia. Another reason could be that women are less likely than men to feel welcome on Wikipedia. Yet another reason could be women don’t disclose their gender as much as their counterparts. In this project, I have used dataset from the UCI Machine Learning repository. It contains 4746 instances each containing 21 attributes with no missing values. Each instance is one person’s unique edit to the Wikipedia page. I have separated the dataset into two separate datasets one for gender selected as unknown and anpther for male and female gender. After identifying the best fit predictive model, I used the model in the dataset with gender value selected as unknown and predicted the gender of those editors whose gender was given as unknown in this dataset. The prediction aligns with my assumption that women don’t disclose their gender as much as their counterparts. Because, 1,555 of 1,601 unknown gender have been predicted as female gender after the prediction.